Directory macro generate struct and default no-argument initialization.
The origin generate class rather than struct, which may influence performance.
The origin also require user to assign the var manually, but now the non-argument initializer is called.
For the case of argument required, user cannot use the `directory` macro and the class or struct must be write by user.